# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/repos/discussions": {
            "get": {
                "description": "get discussions from GitHub",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GitHub"
                ],
                "summary": "get discussions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search by",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.Discussion"
                            }
                        }
                    }
                }
            }
        },
        "/repos/info": {
            "get": {
                "description": "get repo info from GitHub",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GitHub"
                ],
                "summary": "get repo info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.Repo"
                        }
                    }
                }
            }
        },
        "/repos/issues": {
            "get": {
                "description": "get issues from GitHub",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GitHub"
                ],
                "summary": "get issues",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search by",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.Issue"
                            }
                        }
                    }
                }
            }
        },
        "/safeline/count": {
            "get": {
                "description": "get installer count",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Safeline"
                ],
                "summary": "get installer count",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.InstallerCount"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "service.Category": {
            "type": "object",
            "properties": {
                "emoji": {
                    "type": "string"
                },
                "emoji_html": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.Discussion": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/service.User"
                },
                "category": {
                    "$ref": "#/definitions/service.Category"
                },
                "comment_count": {
                    "type": "integer"
                },
                "comment_users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.User"
                    }
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_answered": {
                    "type": "boolean"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Label"
                    }
                },
                "thumbs_up": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "upvote_count": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "service.InstallerCount": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                }
            }
        },
        "service.Issue": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/service.User"
                },
                "comment_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Label"
                    }
                },
                "thumbs_up": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "service.Label": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.Repo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "star_count": {
                    "type": "integer"
                }
            }
        },
        "service.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                }
            }
        }
    }
}